refactor(plugins): Split plugins-api from plugins-spring-api
