+++ kork-expressions/kork-expressions.gradle
@@ -10,6 +10,7 @@ dependencies {
   api(platform(project(":spinnaker-dependencies")))
 
   api project(":kork-plugins-api")
+  api project(":kork-exceptions")
   api "com.fasterxml.jackson.core:jackson-databind"
   api "org.springframework:spring-context"
   api "org.slf4j:slf4j-api"
