+++ kork-plugins/kork-plugins.gradle
@@ -6,11 +6,11 @@ apply from: "$rootDir/gradle/lombok.gradle"
 dependencies {
   api(platform(project(":spinnaker-dependencies")))
   api project(":kork-exceptions")
-  implementation project(":kork-plugins-api")
+  implementation project(":kork-plugins-spring-api")
 
   implementation "org.pf4j:pf4j"
 
-  api "org.springframework.boot:spring-boot-starter-web"
+  implementation "org.springframework.boot:spring-boot-starter-web"
   implementation("com.google.guava:guava")
 
   implementation("com.fasterxml.jackson.dataformat:jackson-dataformat-yaml")
