+++ kork-plugins/src/test/kotlin/com/netflix/spinnaker/kork/plugins/types.kt
@@ -15,7 +15,7 @@
  */
 package com.netflix.spinnaker.kork.plugins
 
-import com.netflix.spinnaker.kork.plugins.api.SpringPlugin
+import com.netflix.spinnaker.kork.plugins.api.spring.SpringPlugin
 import org.pf4j.Extension
 import org.pf4j.PluginWrapper
 import org.springframework.beans.factory.annotation.Autowired
